feat: adding contextual relevance to language translation
- This is done in order to avoid ambiguity in language translation when one word in a non english language can mean the same thing in english.
eg: `vendakka` in malayalam can be translated to `okra` or `ladies finger`. Both are the same. Contextual relevance is added so that the final database will only contain one instance of either okra or ladies finger but not both. This is done ion order to maintain consistency while retreiving the data in the further steps and to also give proper reporting later on in the stages.

+++ internal/models/models.go
@@ -38,10 +38,10 @@ type InventoryData struct {
 }
 type SalesData struct {
 	Item                   string  `json:"item"`
-	Quantity               float32 `json:"quantity"`
+	Quantity               float64 `json:"quantity"`
 	Unit                   string  `json:"unit"`
-	RetailPricePerQuantity float32 `json:"retail_price_per_quantity"`
-	TotalSellingPrice      float32 `json:"total_selling_price"`
+	RetailPricePerQuantity float64 `json:"retail_price_per_quantity"`
+	TotalSellingPrice      float64 `json:"total_selling_price"`
 }
 
 // global api response models
